+++ contract-nft/src/whitelist.rs
@@ -17,7 +17,7 @@ impl Contract {
     //Returns `true` if the account id was not in the whitelist before, `false` otherwise.
     //This method can be called only by the smart contract owner.
     #[private]
-    pub fn add_to_whitelist(&mut self, account_id: AccountId) -> bool {
+    pub fn add_to_whitelist(&mut self, account_id: &AccountId) -> bool {
 
         //ensure the function was called by the smart contract owner, else panic
         self.assert_called_by_owner();
